requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
jinja2>=3.1.0
numpy>=1.26.0
joblib>=1.4.0
scikit-learn>=1.5.0
torch>=2.3.0
pytorch-tabnet>=4.1.0
requests>=2.32.0
httpx>=0.27.0
redis>=5.0.0
python-dotenv>=1.0.1
prometheus-client>=0.20.0
python-json-logger>=2.0.7
PyJWT>=2.8.0
boto3>=1.34.0
azure-identity>=1.17.0
azure-mgmt-resource>=23.1.1
google-cloud-resource-manager>=1.14.2
